+++ lib/puree/journal.rb
@@ -1,0 +1,13 @@
+module Puree
+
+  # Journal resource
+  #
+  class Journal < Resource
+
+    def initialize
+      super(:journal)
+    end
+
+  end
+
+end