+++ lib/puree/organisation.rb
@@ -1,0 +1,13 @@
+module Puree
+
+  # Organisation resource
+  #
+  class Organisation < Resource
+
+    def initialize
+      super(:organisation)
+    end
+
+  end
+
+end