+++ lib/puree/person.rb
@@ -1,0 +1,13 @@
+module Puree
+
+  # Person resource
+  #
+  class Person < Resource
+
+    def initialize
+      super(:person)
+    end
+
+  end
+
+end