+++ lib/puree/project.rb
@@ -1,0 +1,13 @@
+module Puree
+
+  # Project resource
+  #
+  class Project < Resource
+
+    def initialize
+      super(:project)
+    end
+
+  end
+
+end