+++ lib/puree/publication.rb
@@ -1,0 +1,13 @@
+module Puree
+
+  # Publication resource
+  #
+  class Publication < Resource
+
+    def initialize
+      super(:publication)
+    end
+
+  end
+
+end