+++ lib/puree/publisher.rb
@@ -1,0 +1,13 @@
+module Puree
+
+  # Publisher resource
+  #
+  class Publisher < Resource
+
+    def initialize
+      super(:publisher)
+    end
+
+  end
+
+end