+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.6.0 - 2016-05-06
+### Added
+- API map factored out of Resource.
+- Collections (journal, organisation, person, project, publication, publisher).
+- Resource classes with HTTParty metadata (journal, organisation, person, project, publication, publisher).
+
 ## 0.5.0 - 2016-04-29
 ### Added
 - Refactoring ctor.